# Reduce allocations performed by the Logger returned from log.With.