@tsenart 2015-04-18
I'm not immediately seeing where the `With` method would be called in a tight loop or even repeatedly. Would you please elucidate me?