@ChrisHines 2015-04-19
@tsenart The problem isn't that `With` would be called in a tight loop, but that it will be used to build a few (several?) layers of context. The current implementation of `With` builds a stack of closures that is  flattened in every call to `Log`. Each step of flattening involves an additional slice allocation. This PR flattens the context once during `With` to avoid incurring the cost for every subsequent call to `Log`.

Here is some benchmark code that I think represents a typical scenario (let me know if you think this is worth adding to this PR).

```go
func BenchmarkTypicalWith(b *testing.B) {
	logger := log.NewDiscardLogger()
	logger = log.With(logger, "lvl", "info")
	logger = log.With(logger, "host", "hostname")
	logger = log.With(logger, "svc", "svcname")
	b.ReportAllocs()
	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		l := log.With(logger, "client", "client-id", "item", "item-id")
		l.Log("step", 1)
		l.Log("step", 2, "param", "arg")
		l.Log("step", 3, "err", errors.New("error text"))
	}
}
```
I get the results shown below which I think are significant, especially in light of Peter's statement in the discussion of PR #16 that he, "... own[s] and operate[s] services that can become CPU bound based on log output ...." In my experience, logging impacts CPU primarily by heap allocations and formatting into the final `[]byte`s for output. The formatting can sometimes be done asynchronously, but the allocations this benchmark measures would directly impact response time for a request.
```
benchmark                old ns/op     new ns/op     delta
BenchmarkTypicalWith     10760         6825          -36.57%

benchmark                old allocs     new allocs     delta
BenchmarkTypicalWith     33             23             -30.30%

benchmark                old bytes     new bytes     delta
BenchmarkTypicalWith     2520          1624          -35.56%
```